requirements.txt
plexapi

requests
pyyaml
simplejson

rdflib
SPARQLWrapper

boto3
minio